+++ FncyWallet/Classes/Model/Asset/FncyChainInfo.swift
