+++ FncyWallet/Classes/Model/Asset/FncyCurrency.swift
