+++ FncyWallet/Classes/Model/Asset/FncyPrice.swift
