+++ FncyWallet/Classes/Model/Asset/NFTItemInfo.swift
