+++ FncyWallet/Classes/Model/Common/FncyWalletResponse.swift
@@ -37,46 +37,5 @@ extension Status : CustomStringConvertible  {
     }
 }
 
-public protocol ResultPresentable : Codable {
-    var resultType: String? { get }
-    var result: ResultInfo? { get }
-}
 
-internal struct ListData<C: Codable>: ResultPresentable {
-    let items: C?
-    let resultType: String?
-    let result: ResultInfo?
-    let wid : Int?
-    
-    init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: ListData<C>.CodingKeys.self)
-        self.resultType = try container.decodeIfPresent(String.self, forKey: .resultType)
-        self.result = try container.decodeIfPresent(ResultInfo.self, forKey: .result)
-        self.items = try container.decodeIfPresent(C.self, forKey: .items)
-        
-        self.wid = try container.decodeIfPresent(Int.self, forKey: .wid)
-    }
-}
-
-// PagingListData<C>
-internal struct PagingListData<C: Codable>: ResultPresentable {
-    let items: C?
-    let paging: Paging?
-    let resultType: String?
-    let result: ResultInfo?
-    init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: PagingListData<C>.CodingKeys.self)
-        self.paging = try container.decodeIfPresent(Paging.self, forKey: .paging)
-        self.resultType = try container.decodeIfPresent(String.self, forKey: PagingListData<C>.CodingKeys.resultType)
-        self.result = try container.decodeIfPresent(ResultInfo.self, forKey: PagingListData<C>.CodingKeys.result)
-        self.items = try container.decodeIfPresent(C.self, forKey: PagingListData<C>.CodingKeys.items)
-    }
-}
-
-public struct Paging: Codable {
-    public let pageNo: Int
-    public let pageSize: Int
-    public let totalCount: Int
-    public let hasMore: Bool
-}
